requirements.txt
streamlit
google-genai
pillow
fal-client
requests